% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_datagovin_server}
\alias{check_datagovin_server}
\title{Utility to check the server of data.gov.in APIs}
\usage{
check_datagovin_server()
}
\value{
Message
}
\description{
Utility to check the server of data.gov.in APIs
}
\examples{
check_datagovin_server()
}
