% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_internet_connection}
\alias{check_internet_connection}
\title{Basic Utility to check internet connection}
\usage{
check_internet_connection()
}
\value{
Message
}
\description{
Basic Utility to check internet connection
}
\examples{
check_internet_connection()
}
