% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{get_api_info}
\alias{get_api_info}
\title{Get information about the API using the API index name}
\usage{
get_api_info(api_index)
}
\arguments{
\item{api_index}{API index name (string) you found using the search functions. You can
also get these from data.gov.in from a specific API page. In the request url,
it is followed by /resource/xxxxxxxx
For getting the relevant fields in the api use get_api_fields}
}
\value{
Data frame with 1 row , API that matches API ID
}
\description{
Get information about the API using the API index name
}
\examples{
  get_api_info("3b01bcb8-0b14-4abf-b6f2-c1bfd384ba69")
}
