% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{get_list_of_org_types}
\alias{get_list_of_org_types}
\title{Get a list of unique organization types}
\usage{
get_list_of_org_types()
}
\value{
character vector of organization types
}
\description{
These include whether a Central/State organization releases the underlying data.
This will be helpful while querying for the right API
}
\examples{
get_list_of_org_types()
}
