% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{get_list_of_organizations}
\alias{get_list_of_organizations}
\title{Get a list of unique organizations}
\usage{
get_list_of_organizations()
}
\value{
character vector of organization
}
\description{
These include which ministry releases the underlying data.
This will be helpful while querying for the right API.
Get a list of unique sectors
}
\examples{
get_list_of_organizations()
}
