% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{get_list_of_sectors}
\alias{get_list_of_sectors}
\title{Get a list of unique sectors}
\usage{
get_list_of_sectors()
}
\value{
character vector of sectors
}
\description{
These include which sector the underlying data belongs to.
This will be helpful while querying for the right API.
}
\examples{
get_list_of_sectors()
}
