% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{get_list_of_sources}
\alias{get_list_of_sources}
\title{Get a list of unique data sources (website names)}
\usage{
get_list_of_sources()
}
\value{
character vector of sectors
}
\description{
These include which websites host the underlying data.
This will be helpful while querying for the right API.
}
\examples{
get_list_of_sources()
}
