% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{get_recently_created_api}
\alias{get_recently_created_api}
\title{Get a data.frame of recently created APIs}
\usage{
get_recently_created_api(N = 20)
}
\arguments{
\item{N}{Number of APIs to return}
}
\value{
Get data.frame of N most recently created API sorted by date of creation
}
\description{
This will return a data.frame of recently created APIs
}
\examples{
recently_created_api<-get_recently_updated_api()
}
