% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{get_recently_updated_api}
\alias{get_recently_updated_api}
\title{Get a data.frame of recently updated APIs}
\usage{
get_recently_updated_api(N = 20)
}
\arguments{
\item{N}{Number of APIs to return}
}
\value{
Get data.frame of N most recently updated API sorted by date of update
}
\description{
This will return a data.frame of recently updated APIs
}
\examples{
recently_updated_api<-get_recently_updated_api()
}
