% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_api_details}
\alias{import_api_details}
\title{Utility to import RDS files with API details from Github}
\usage{
import_api_details()
}
\value{
file (data.frame)
}
\description{
Utility to import RDS files with API details from Github
}
\examples{
api_details<-import_api_details()
}
