% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_field_details}
\alias{import_field_details}
\title{Utility to import updated RDS files with API field details from Github}
\usage{
import_field_details()
}
\value{
file (data.frame)
}
\description{
Utility to import updated RDS files with API field details from Github
}
\examples{
api_field_details<-import_field_details()
}
