% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_rds_from_github}
\alias{read_rds_from_github}
\title{Utility to import RDS files from Github}
\usage{
read_rds_from_github(url)
}
\arguments{
\item{url}{url of the raw girhub file}
}
\value{
file (data frame)
}
\description{
Utility to import RDS files from Github
}
\examples{
read_rds_from_github(url)
}
