% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{register_api_key}
\alias{register_api_key}
\title{Registering/Validating User API key}
\usage{
register_api_key(user_api_key, show_key = TRUE)
}
\arguments{
\item{user_api_key}{API Key obtained from data.gov.in}

\item{show_key}{Whether to API show key in messages}
}
\value{
Success/Faliure of API key validation
}
\description{
Obtain an API key from data.gov.in by registering on the platform. You can find
it on your "My Account page after logging in.
}
\examples{
\dontrun{
register_api_key(api_key=xxx,show_key=FALSE)
}
}
