% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{search_api_by_created_date}
\alias{search_api_by_created_date}
\title{Search API by Created Date}
\usage{
search_api_by_created_date(date = Sys.Date())
}
\arguments{
\item{date}{A Date object corresponding to chosen date of creation}
}
\value{
data.frame of API details filtered by date of creation
}
\description{
Returns APIs which were created on a specific date.
This function takes in a Date object as its input. Use ymd/mdy/dmy functions
from the package lubridate to make this easier.
}
\examples{
filtered_api<-search_api_by_created_date(date=as.Date("2021-12-20"))
}
