% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{search_api_by_description}
\alias{search_api_by_description}
\title{Search API by description}
\usage{
search_api_by_description(description_contains = "")
}
\arguments{
\item{description_contains}{string to match in the description of the API (case insensitive)}
}
\value{
data.frame of API details filtered by descriptions that match the string provided
}
\description{
Search API by description
}
\examples{
filtered_api<-search_api_by_description(description_contains="Air Quality")
}
