% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{search_api_by_org_type}
\alias{search_api_by_org_type}
\title{Search API by organization type}
\usage{
search_api_by_org_type(org_type_contains = "")
}
\arguments{
\item{org_type_contains}{string to match in the organization type of the API (case insensitive)}
}
\value{
data.frame of API details filtered by organization types that match the string provided
}
\description{
Search API by organization type
}
\examples{
  get_list_of_org_types()
 filtered_api<-search_api_by_org_type(org_type_contains="Central")
}
