% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{search_api_by_organization}
\alias{search_api_by_organization}
\title{Search API by organization name}
\usage{
search_api_by_organization(organization_name_contains = "")
}
\arguments{
\item{organization_name_contains}{string to match in the organization type of the API (case insensitive)}
}
\value{
data.frame of API details filtered by organization names (Ministries) that match the string provided
}
\description{
Search API by organization name
}
\examples{
  get_list_of_organizations()
filtered_api<-search_api_by_organization(organization_name_contains="Agriculture")

}
