% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{search_api_by_sector}
\alias{search_api_by_sector}
\title{Search API by Sector}
\usage{
search_api_by_sector(sector_name_contains = "")
}
\arguments{
\item{sector_name_contains}{string to match in the organization type of the API (case insensitive)}
}
\value{
data.frame of API details filtered by Sectors that match the string provided
}
\description{
Search API by Sector
}
\examples{
get_list_of_sectors()
filtered_api<-search_api_by_sector(sector_name_contains="Consumer")
}
