% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{search_api_by_source}
\alias{search_api_by_source}
\title{Search API by Data Source}
\usage{
search_api_by_source(source_name_contains = "")
}
\arguments{
\item{source_name_contains}{string to match in the organization type of the API (case insensitive)}
}
\value{
data.frame of API details filtered by Sectors that match the string provided
}
\description{
Search API by Data Source
}
\examples{
get_list_of_sources()
filtered_api<-search_api_by_source(source_name_contains="gov")
}
