% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_func.R
\name{search_api_by_title}
\alias{search_api_by_title}
\title{Search API by title}
\usage{
search_api_by_title(title_contains = "")
}
\arguments{
\item{title_contains}{string to match in the title of the API (case insensitive)}
}
\value{
data.frame of API details filtered by titles that match the string provided
}
\description{
Search API by title
}
\examples{
filtered_api<-search_api_by_title(title_contains="Air Quality")
}
